requirements.txt
openai
beautifulsoup4
requests
praw
tweepy
python-wordpress-xmlrpc
python-dotenv
schedule
lxml
